Re-added ILOC files from Blackboard

+++ block-4/pp/iloc/eval/Memory.java
@@ -5,12 +5,12 @@ import java.util.List;
 
 /** Simulated memory. */
 public class Memory {
-	private final List<Integer> mem = new ArrayList<>(0);
+	private final List<Byte> mem = new ArrayList<>(0);
 
 	/** Sets a location in memory to a given value. */
-	public void set(int loc, int value) {
+	public void set(int loc, byte value) {
 		while (loc >= this.mem.size()) {
-			this.mem.add(0);
+			this.mem.add((byte) 0);
 		}
 		this.mem.set(loc, value);
 	}
@@ -19,9 +19,9 @@ public class Memory {
 	 * Returns the value at a given memory location.
 	 * The value is 0 if the location was never accessed before. 
 	 */
-	public int get(int loc) {
+	public byte get(int loc) {
 		while (loc >= this.mem.size()) {
-			this.mem.add(0);
+			this.mem.add((byte) 0);
 		}
 		return this.mem.get(loc);
 	}
